fix: price showing and ens action
