+++ pages/search.tsx
@@ -22,7 +22,10 @@ const SearchPage: NextPage = () => {
   const [ppImageUrl, setPpImageUrl] = useState("");
 
   useEffect(() => {
-    if (!identity || !address) return;
+    if (!identity || !address) {
+      setIsOwner(false);
+      return;
+    }
     setIsOwner(identity.ownerAddress === formatHexString(address));
   }, [identity, address]);
