+++ styles/components/identityCard.module.css
@@ -202,7 +202,7 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-  gap: 12px;
+  gap: 8px;
   cursor: pointer;
   font-size: 13px;
   color: #454545;
@@ -217,7 +217,7 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-  gap: 12px;
+  gap: 8px;
 }
 
 .evmAddrBtn {
